# contracts/openapi/riverhog.v1.yaml
openapi: 3.1.0
info:
  title: RIVERHOG API
  version: 1.0.0
  summary: MVP API for hot/cold archival storage
  description: Authoritative archive resources are durable across service restarts unless a specific contract explicitly expires or releases them.
servers:
  - url: /v1
paths:
...
  /images/{image_id}/disc-rebuild:
    get:
      summary: Get the latest disc rebuild archive restore for one image
      description: Returns the latest `disc_rebuild` archive restore for one finalized image, including expired or completed restores.
      operationId: getImageRebuildSession
      parameters:
        - in: path
          name: image_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Archive restore summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveRestore"
        default:
          $ref: "#/components/responses/Error"
  /archive-restores:
    get:
      summary: List archive restores
      description: Returns a paged inventory of archive-backed fetch materialization and disc rebuild restores, with filters for lifecycle scope, type, state, collection, and image.
      operationId: listArchiveRestores
      parameters:
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: per_page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - in: query
          name: sort
          required: false
          schema:
            type: string
            enum:
              [created_at, id, type, state, ready_at, expires_at]
            default: created_at
        - in: query
          name: order
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - in: query
          name: terminal
          required: false
          schema:
            type: string
            enum: [active, terminal, all]
            default: all
        - in: query
          name: type
          required: false
          schema:
            oneOf:
              - type: string
                enum: [fetch_materialization, disc_rebuild]
              - type: "null"
        - in: query
          name: state
          required: false
          schema:
            oneOf:
              - type: string
                enum: [requested, ready, paused, expired, completed, failed, canceled]
              - type: "null"
        - in: query
          name: collection
          required: false
          schema:
            oneOf:
              - type: string
              - type: "null"
        - in: query
          name: image
          required: false
          schema:
            oneOf:
              - type: string
              - type: "null"
      responses:
        "200":
          description: Paged archive restore inventory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveRestoreList"
        default:
          $ref: "#/components/responses/Error"
  /archive-restores/{archive_restore_id}:
    get:
      summary: Get one archive restore
      description: Returns one archive-backed fetch materialization or disc rebuild archive restore by durable archive restore id.
      operationId: getArchiveRestore
      parameters:
        - in: path
          name: archive_restore_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Archive restore summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveRestore"
        default:
          $ref: "#/components/responses/Error"
  /archive-restores/{archive_restore_id}/complete:
    post:
      summary: Complete one ready archive restore
      description: Marks one ready archive restore complete and records cleanup or lifecycle handoff for restored Standard-storage data.
      operationId: completeArchiveRestore
      parameters:
        - in: path
          name: archive_restore_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Completed archive restore summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveRestore"
        default:
          $ref: "#/components/responses/Error"
  /archive-restores/{archive_restore_id}/cancel:
    post:
      summary: Cancel one active fetch materialization archive restore
      description: Cancels one active fetch-scoped fetch materialization archive restore by durable archive restore id and stops automatic materialization. Disc rebuild restores are paused and resumed instead of canceled.
      operationId: cancelArchiveRestore
      parameters:
        - in: path
          name: archive_restore_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Canceled archive restore summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveRestore"
        default:
          $ref: "#/components/responses/Error"
  /archive-restores/{archive_restore_id}/pause:
    post:
      summary: Pause one active disc rebuild archive restore
      description: Pauses one active `disc_rebuild` archive restore and schedules daily operator reminders until it is resumed.
      operationId: pauseArchiveRestore
      parameters:
        - in: path
          name: archive_restore_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Paused archive restore summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveRestore"
        default:
          $ref: "#/components/responses/Error"
  /archive-restores/{archive_restore_id}/resume:
    post:
      summary: Resume one paused disc rebuild archive restore
      description: Resumes one paused `disc_rebuild` archive restore from its previous restore state.
      operationId: resumeArchiveRestore
      parameters:
        - in: path
          name: archive_restore_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Resumed archive restore summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveRestore"
        default:
          $ref: "#/components/responses/Error"
  /archive-restores/{archive_restore_id}/images/{image_id}/iso:
    get:
      summary: Download one rebuilt archive-restore ISO
      description: Streams rebuilt ISO bytes for one image in a ready `disc_rebuild` archive restore.
      operationId: getRecoveredIso
      parameters:
        - in: path
          name: archive_restore_id
          required: true
          schema: { type: string }
        - in: path
          name: image_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Rebuilt ISO bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        default:
          $ref: "#/components/responses/Error"
...
components:
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    Archive:
      type: object
      required:
        [state, object_path, stored_bytes, backend, storage_class, last_uploaded_at, last_verified_at, failure]
      properties:
        state:
          type: string
          enum: [pending, uploading, uploaded, retrying, failed]
        object_path:
          oneOf:
            - type: string
            - type: "null"
        stored_bytes:
          oneOf:
            - type: integer
            - type: "null"
        backend:
          oneOf:
            - type: string
            - type: "null"
        storage_class:
          oneOf:
            - type: string
            - type: "null"
        last_uploaded_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        last_verified_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        failure:
          oneOf:
            - type: string
            - type: "null"
    CollectionManifestProofState:
      type: string
      enum: [pending, uploaded, failed]
    CollectionManifest:
      type: object
      required: [object_path, sha256, ots_object_path, ots_state]
      properties:
        object_path:
          oneOf:
            - type: string
            - type: "null"
        sha256:
          oneOf:
            - type: string
              pattern: "^[0-9a-f]{64}$"
            - type: "null"
        ots_object_path:
          oneOf:
            - type: string
            - type: "null"
        ots_state:
          $ref: "#/components/schemas/CollectionManifestProofState"
...
    ArchiveRestoreList:
      type: object
      required:
        [page, per_page, total, pages, sort, order, terminal, type, state, collection, image, restores]
      properties:
        page:
          type: integer
          minimum: 1
        per_page:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        pages:
          type: integer
          minimum: 0
        sort:
          type: string
          enum: [created_at, id, type, state, ready_at, expires_at]
        order:
          type: string
          enum: [asc, desc]
        terminal:
          type: string
          enum: [active, terminal, all]
        type:
          oneOf:
            - type: string
              enum: [fetch_materialization, disc_rebuild]
            - type: "null"
        state:
          oneOf:
            - type: string
              enum: [requested, ready, paused, expired, completed, failed, canceled]
            - type: "null"
        collection:
          oneOf:
            - type: string
            - type: "null"
        image:
          oneOf:
            - type: string
            - type: "null"
        restores:
          type: array
          items:
            $ref: "#/components/schemas/ArchiveRestore"
    ArchiveRestore:
      type: object
      required:
        [id, type, state, created_at, warnings, notification, progress, collections, images]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [fetch_materialization, disc_rebuild]
        state:
          type: string
          enum: [requested, ready, paused, expired, completed, failed, canceled]
        created_at:
          type: string
          format: date-time
        requested_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        ready_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        expires_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        completed_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        canceled_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        paused_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        paused_from_state:
          oneOf:
            - type: string
            - type: "null"
        paths:
          oneOf:
            - type: array
              items:
                type: string
            - type: "null"
        latest_message:
          oneOf:
            - type: string
            - type: "null"
        warnings:
          type: array
          items:
            type: string
        notification:
          $ref: "#/components/schemas/RecoveryNotificationStatus"
        progress:
          $ref: "#/components/schemas/ArchiveRestoreProgress"
        collections:
          type: array
          items:
            $ref: "#/components/schemas/ArchiveRestoreCollection"
        images:
          type: array
          items:
            $ref: "#/components/schemas/ArchiveRestoreImage"
    RecoveryNotificationStatus:
      type: object
      required:
        [
          webhook_configured,
          reminder_count,
          next_reminder_at,
          last_notified_at,
          failure_count,
          last_failure_at,
          last_failure,
        ]
      properties:
        webhook_configured:
          type: boolean
        reminder_count:
          type: integer
        next_reminder_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        last_notified_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        failure_count:
          type: integer
          minimum: 0
        last_failure_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        last_failure:
          oneOf:
            - type: string
            - type: "null"
    ArchiveRestoreProgress:
      type: object
      required: [archive_verification, extraction, materialization]
      properties:
        archive_verification:
          type: string
          enum: [pending, in_progress, completed, failed]
        extraction:
          type: string
          enum: [pending, in_progress, completed, failed]
        materialization:
          type: string
          enum: [pending, in_progress, completed, failed]
    ArchiveRestoreImage:
      type: object
      required: [id, filename, collection_ids, rebuild_state]
      properties:
        id:
          type: string
        filename:
          type: string
        collection_ids:
          type: array
          items: { type: string }
        rebuild_state:
          type: string
          enum: [pending, restoring_collections, rebuilding, ready, paused, failed, canceled]
    ArchiveRestoreCollection:
      type: object
      required: [id, archive, collection_manifest, stored_bytes]
      properties:
        id:
          type: string
        archive:
          $ref: "#/components/schemas/Archive"
        collection_manifest:
          $ref: "#/components/schemas/CollectionManifest"
        stored_bytes:
          type: integer
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }